improved and cleared the code
