+++ document-fields.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import { FormField } from '@keystone-6/fields-document/component-blocks'
+import { ImageUploader } from './components/ImageUploader'
+import { TImageValue } from './components/ImageUploader/types'
+
+export type TImageOptions = {
+  listKey: string
+}
+
+export const image = ({ listKey }: TImageOptions): FormField<TImageValue, TImageOptions> => {
+  return {
+    kind: 'form',
+    Input({ value, onChange }) {
+      return <ImageUploader listKey={listKey} defaultValue={value} mode="edit" onChange={onChange} />
+    },
+    options: { listKey },
+    defaultValue: null,
+    validate(value) {
+      return typeof value === 'object'
+    },
+  }
+}
