+++ fields.ts
@@ -2,7 +2,7 @@ import { integer, text, timestamp, select, image, relationship } from '@keystone
 import { document } from '@keystone-6/fields-document'
 import slugify from 'slugify'
 import { mainConfig } from './config'
-import { componentBlocks } from './plugins/component-blocks'
+import { componentBlocks } from './component-blocks'
 
 export const slugField = text({
   defaultValue: '',
@@ -109,7 +109,7 @@ export const contentField = document({
   links: true,
   dividers: true,
   ui: {
-    views: require.resolve('./plugins/component-blocks'),
+    views: require.resolve('./component-blocks'),
   },
   componentBlocks,
 })
