+++ helpers/importJson/constants.ts
@@ -1,0 +1,5 @@
+export const IMPORT_DIR = './import_data'
+export const DEFAULT_PASSWORD = 'qwer1234'
+export const DEFAULT_WIDTH = 800
+export const DEFAULT_HEIGHT = 600
+export const DEFAULT_SIZE = 1024
