+++ helpers/transformFilename.ts
@@ -2,9 +2,9 @@ import cuid from 'cuid'
 import slugify from 'slugify'
 import { parseFilename } from './parseFilename'
 
-export const transformFilename = (fullFilename: string) => {
-  const { filename } = parseFilename(fullFilename)
-  const safeFilename = slugify(filename, { replacement: '-' })
+export const transformFilename = (filename: string) => {
+  const { name } = parseFilename(filename)
+  const safeFilename = slugify(name, { replacement: '-' })
   const id = cuid()
 
   return `${safeFilename}-${id}`
