--- plugins/types.ts
@@ -1,12 +1,0 @@
-export type TImageFieldData = {
-  id: string
-  image?: {
-    url: string
-  }
-}
-
-export type TImageFieldValue = TImageFieldData | null
-
-export type TImageFieldOptions = {
-  listKey: string
-}
