+++ tsconfig.json
@@ -7,5 +7,9 @@
     "noEmit": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true
-  }
+  },
+  "exclude": [
+    "node_modules",
+    "./.keystone",
+  ],
 }
